Ares BMP screenshot image processing changes.
- Moved ProcessScreenshots method from GameLobbyBase to GameInProgressWindow, as that is where other filesystem operations are performed before/after starting game.
- Removed ProcessScreenshots setting in ClientConfiguration, the feature is now always enabled on AresRelease configuration of the client.
